requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
pydantic==1.10.14
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
qrcode==7.4.2
Pillow==10.3.0
pymysql==1.1.0

